requirements.txt
streamlit
pdfplumber
python-dotenv
requests
spacy